requirements.txt
#   pip install -r streamlit-oauth


streamlit
streamlit-chat
langchainhub
wikipedia
duckduckgo-search
openai
langchain
langchain-openai
langchain-community
python-dotenv

# 캘린더
streamlit_calendar

# 로그인 관련
streamlit-oauth

# 구글 권한 관련
google-api-python-client
google-auth
google-auth-oauthlib

# 번역 라이브러리 
deep-translator

# pdf 생성 라이브러리
aspose-pdf 
